+++ client/Admin/index.js
@@ -1,3 +1,3 @@
 //Admin components
 export {default as AdminMain} from './main'
-export {default as Orders} from './Orders'
+export {default as AllOrders} from './AllOrders'
